Ajustado Busca por ID - Falta tratar erro de caracter. Ajustado Busca por categoria Ajustado Busca por localidade Ajustado Busca por Dias de trabalho - Falta tratar erro de caracter.

+++ PI_ALESSANDRA/src/pi_alessandra/DadosVagas.java
@@ -5,6 +5,6 @@ public class DadosVagas {
     public String categoria;
     public String localVagas;   
     public double salarioDiaristas;
-    public byte diasTrabalhados;
+    public int diasTrabalhados;
     public String detalhesVagas;
 }
